Add libabigail consistency test framework
To be able to use libabigail's abixml as a stored format for a
binary's ABI information in place of the binary itself. The
information in the ABIXML file needs to faithfully represent the ABI
of the file being tested. This adds a test framework that verifies
that the target shared libraries do have their information represented
properly.

Signed-off-by: Ben Woodard <woodard@redhat.com>