Reorganize to facilitate future testing
Reorganized the directory structure to separate test cases and tools from the
abichecking tools that they are supposed to test. Currently there are no tests
for smeagol but as soon as smeagol is ready to be tested these test cases are
some that both libabigail and smeagol need to be able to perform correctly.