Wire up exception tests & finish reorg
The exception ABI tests were in the repo but they were not wired up yet.
This patch wires up the tests so that they run properly.

While looking at wiring up the exception tests, I realized that there
was a difference between the little tests which demonstrate the ABI problems
and the programs used to evaluate the tests. I moved things around so that
the tests that demonstrate ABI breaks are in the tests area while the code
which shows problems in the tested ABI break detection tools are in their
own subdir.