Modify tests to handle non-tpm owned dirs
Those scripts were made for testing the distro. However, I wanted to
expand my testing to include spack repos. This required some modification to
the scripts so that they didn't bother testing libraries which didn't have
debuginfo and so that they didn't look for the source of the files in a RPM.