Add new backcall test with weak symbols
While trying to make a new test to address the backcall problem and
avoid the NOTYPE problem, I used weak symbols. In this test the strong
symbol in the app overrides the weak symbol in the library. Since they
don't have the same ABI, the program breaks. Libabigail notices the
ABI change when diffing the libraries but it fails to recognize the
ABI compatibility problem when doing an abicompat.